+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `extend` method, for easier extending
 - `parent` property, for accessing parent error
+- `normalize` method, for easy normalization of invalid values
 ### Changed
 - README cleanup