+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [6.0.0-beta.1] - 2023-02-28
+### Added
+- `normalize` 2nd parameter, to choose normalization mode
+
 ## [5.0.0] - 2023-02-23
 ### Added
 - `extend` method, for easier extending